added test for won method
