--- spec/tic_tac_toe_rspec.rb
@@ -1,3 +1,0 @@
-require_relative '../lib/board.rb'
-require_relative '../lib/play_game.rb'
-require_relative '../bin/main.rb'
